stuff was added lol
Made a few modifications to the User, MovieList,  MoviePage, AccountPage, and SearchPage classes.
For User:
Added a function to add a movie to the users favorite movies.

For MovieList:
Added a toString function and print function.
Made it to where the year a movie was released is added to the title of the movie i.e. Dune (2021).
Made it to where the movies in the MovieList are sorted alphabetically

For MoviePage:
Added a '<3' button so the user can add it to their favorite movies.

For AccountPage:
Added functionality to the display of custom MovieLists to display the MovieList name in the button. Buttons are only created if a MovieList is in the list of MovieLists.
Added basically the same functionality to the Favorite Movies display.
Clicking on the buttons in the favorite movies opens a dialog with the movie page for that movie.

For SearchPage:
Did a little clean up
Oh and the biggie. For most of all the backend classes they now implement Serializable so they can be stored in the  "databases"